+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added *Scene Changed* auto execution option.
 - Added full ADSR (Attack, Delay, Sustain, Release) envelope to Evaluate MIDI Track node.
 - Added velocity sensitivity to Evaluate MIDI Track node.
+- Added Container Type advanced option to object instancer node.
 
 ### Fixed